+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - `LocalDate`:
   - `getMonth()` now returns a `Month` enum; use `getMonthValue()` if you want the month number
+- `LocalDateTime`:
+  - `getMonth()` now returns a `Month` enum; use `getMonthValue()` if you want the month number
 
 ## [0.6.5](https://github.com/brick/date-time/releases/tag/0.6.5) - 2024-06-19